requirements.txt
PyQt5>=5.6
pyserial~=3.4